修复网络通信Bug

+++ fastrpc-core/src/main/java/com/fastrpc/remoting/message/Message.java
@@ -16,7 +16,13 @@ import java.util.concurrent.ConcurrentHashMap;
 @Data
 public abstract class Message implements Serializable {
 
+    /**
+     * 用于服务器向客户端发送消息,客户端知道服务器返回的是哪次通信的结果
+     */
     private int seqId;
+    /**
+     * 用于反射创建对应的消息
+     */
     private byte messageType;
 
     private static final Map<Byte,Class<? extends Message>> messageTypeMap=new ConcurrentHashMap<>();
@@ -25,7 +31,12 @@ public abstract class Message implements Serializable {
     {
         return  messageTypeMap.get(messageType);
     }
-    public abstract int getMessageType();
+
+    /**
+     * 获得消息类型
+     * @return 返回消息类型
+     */
+    public abstract byte getMessageType();
 
     /**
      * 请求类型
